+++ frontend/src/components/jobs/job-header.tsx
@@ -1,66 +1,20 @@
 // frontend/src/components/jobs/job-header.tsx
-import { Text } from "@mantine/core";
 import {
-  IconMapPin,
-  IconCalendar,
   IconBriefcase,
   IconBuilding,
+  IconCalendar,
   IconId,
+  IconMapPin,
 } from "@tabler/icons-react";
-import { Job, WORKING_RIGHTS, WorkingRight } from "@/types/job";
-import { formatCapString, getTimeAgo } from "@/lib/utils";
+import { Job } from "@/types/job";
+import { formatCapString, formatWorkingRights, getTimeAgo } from "@/lib/utils";
 import Link from "next/link";
 import CompanyLogo from "@/components/jobs/company-logo";
+import { InfoTag } from "@/components/jobs/info-tag";
 
 interface JobHeaderProps {
   job: Job;
 }
-
-interface InfoTagProps {
-  icon: React.ReactNode;
-  text: string;
-}
-
-const formatWorkingRights = (rights: WorkingRight[]): string => {
-  // If all rights are present, return "Any"
-  if (rights.length === WORKING_RIGHTS.length) {
-    return "Any Working Rights";
-  }
-
-  // Check for AUS and NZ Citizens/PR combination
-  const hasAus = rights.includes("AUS_CITIZEN_PR");
-  const hasNz = rights.includes("NZ_CITIZEN_PR");
-  if (hasAus && hasNz) {
-    return "AUS & NZ Citizen/PR";
-  }
-
-  // Format remaining cases
-  return rights
-    .map((right) => {
-      switch (right) {
-        case "AUS_CITIZEN_PR":
-          return "AUS Citizen/PR";
-        case "NZ_CITIZEN_PR":
-          return "NZ Citizen/PR";
-        case "INTERNATIONAL":
-          return "International";
-        case "OTHER_RIGHTS":
-          return "Other";
-        default:
-          return formatCapString(right);
-      }
-    })
-    .join(", ");
-};
-
-const InfoTag = ({ icon, text }: InfoTagProps) => (
-  <div className="inline-flex items-center gap-1 bg-selected py-1 px-2 rounded-lg">
-    {icon}
-    <Text size="sm" className="whitespace-normal">
-      {text}
-    </Text>
-  </div>
-);
 
 export default function JobHeader({ job }: JobHeaderProps) {
   return (
